+++ build.sh
@@ -12,7 +12,6 @@ mkdir -p build-chrome
 
 rollup ts-out/background.js --format iife --name 'background' --file build-chrome/background.js
 rollup ts-out/content.js --format iife --name 'background' --file build-chrome/content.js
-rollup ts-out/viewer.js --format iife --name 'background' --file build-chrome/viewer.js
 cp src/viewer.css build-chrome/viewer.css
 cp src/manifest.chrome.json build-chrome/manifest.json
 cp license.txt build-chrome/license.txt
@@ -31,7 +30,6 @@ mkdir -p build-firefox
 
 rollup ts-out/background.js --format iife --name 'background' --file build-firefox/background.js
 rollup ts-out/content.js --format iife --name 'background' --file build-firefox/content.js
-rollup ts-out/viewer.js --format iife --name 'background' --file build-firefox/viewer.js
 cp src/viewer.css build-firefox/viewer.css
 cp src/manifest.firefox.json build-firefox/manifest.json
 cp license.txt build-firefox/license.txt