+++ buildWindows.ps1
@@ -7,7 +7,6 @@ mkdir -p build-chrome
 
 yarn rollup ts-out/background.js --format iife --name 'background' --file build-chrome/background.js
 yarn rollup ts-out/content.js --format iife --name 'background' --file build-chrome/content.js
-yarn rollup ts-out/viewer.js --format iife --name 'background' --file build-chrome/viewer.js
 cp src/viewer.css build-chrome/viewer.css
 cp src/manifest.json build-chrome/manifest.json
 cp license.txt build-chrome/license.txt
@@ -25,7 +24,6 @@ mkdir -p build-firefox
 
 yarn rollup ts-out/background.js --format iife --name 'background' --file build-firefox/background.js
 yarn rollup ts-out/content.js --format iife --name 'background' --file build-firefox/content.js
-yarn rollup ts-out/viewer.js --format iife --name 'background' --file build-firefox/viewer.js
 cp src/viewer.css build-firefox/viewer.css
 cp src/manifest.json build-firefox/manifest.json
 cp license.txt build-firefox/license.txt