+++ src/content.ts
@@ -3,23 +3,11 @@ import { errorPage, jsonToHTML } from "./jsonformatter";
 import { installCollapseEventListeners } from "./collapse";
 import { safeStringEncodeNums } from "./safe-encode-numbers";
 
-function setJsonAsGlobalVariable(this: any, jsonObj: any) {
-  const script = document.createElement("script");
-  script.text = `Object.defineProperty(window, 'data', { value: ${JSON.stringify(
-    jsonObj
-  )}, writable: false, configurable: false });`;
-  document.documentElement.appendChild(script);
-
-  // log info message
-  // with this queueMicrotask user can not see source file information in log
-  queueMicrotask(() => console.log('JSON is exposed as a global variable called "data"'));
-}
-
 /**
  * This script runs on every page. It communicates with the background script
  * to help decide whether to treat the contents of the page as JSON.
  */
-chrome.runtime.sendMessage({}, (response: boolean) => {
+chrome.runtime.sendMessage("jsonview-is-json", (response: boolean) => {
   if (!response) {
     return;
   }
@@ -53,5 +41,4 @@ chrome.runtime.sendMessage({}, (response: boolean) => {
 
   document.documentElement.innerHTML = outputDoc;
   installCollapseEventListeners();
-  setJsonAsGlobalVariable(jsonObj);
 });